This commit was manufactured by cvs2svn to create branch 'bf-092-database-1-final-Jan-19-2003'.
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/branches/bf-092-database-1-final-Jan-19-2003@1369 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ bogofilter/qp.h
@@ -12,6 +12,6 @@ NAME:
 
 #include "common.h"
 
-int qp_decode(byte *buff, size_t size);
+size_t qp_decode(byte *buff, size_t size);
 
 #endif	/* HAVE_QP_H */
